# openapi.yaml
openapi: 3.0.0
paths:
...
  /user/order:
    post:
      operationId: OrderController_createOrder
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/OrderResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      security: &ref_4
        - bearer: []
      summary: Create new order
      tags: &ref_5
        - Order
  /user/order/{id}:
    get:
      operationId: OrderController_getOrder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/OrderResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      security: *ref_4
      summary: Get order by ID
      tags: *ref_5
    delete:
      operationId: OrderController_cancelOrder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelOrderDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/OrderResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      security: *ref_4
      summary: Cancel order
      tags: *ref_5
  /user/order/{id}/confirm:
    patch:
      operationId: OrderController_confirmOrder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/OrderResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      security: *ref_4
      summary: Confirm order (shop owner)
      tags: *ref_5
  /user/order/{id}/pay:
    patch:
      operationId: OrderController_markOrderAsPaid
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkOrderAsPaidDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/OrderResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      security: *ref_4
      summary: Mark order as paid (payment confirmation)
      tags: *ref_5
  /user/order/{id}/ship:
    patch:
      operationId: OrderController_shipOrder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShipOrderDto'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    $ref: '#/components/schemas/OrderResponseDto'
                    description: Response data
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
      security: *ref_4
      summary: Ship order (shop owner)
      tags: *ref_5
...
  /user/order/customer/{customerId}:
    get:
      operationId: OrderController_getOrdersByCustomer
      parameters:
        - name: customerId
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
        - name: status
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful paginated response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderResponseDto'
                    description: Array of data items
                  paging:
                    type: object
                    description: Pagination metadata
                    properties:
                      total:
                        type: number
                        example: 100
                        description: Total number of items
                      page:
                        type: number
                        example: 1
                        description: Current page number
                      limit:
                        type: number
                        example: 10
                        description: Items per page
                      totalPages:
                        type: number
                        example: 10
                        description: Total number of pages
                    required:
                      - total
                      - page
                      - limit
                      - totalPages
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
                  - data
                  - paging
      security: *ref_4
      summary: Get orders by customer
      tags: *ref_5
  /user/order/shop/{shopId}:
    get:
      operationId: OrderController_getOrdersByShop
      parameters:
        - name: shopId
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
        - name: status
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful paginated response
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 200
                    description: HTTP status code
                  message:
                    type: string
                    example: OK
                    description: Response message
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderResponseDto'
                    description: Array of data items
                  paging:
                    type: object
                    description: Pagination metadata
                    properties:
                      total:
                        type: number
                        example: 100
                        description: Total number of items
                      page:
                        type: number
                        example: 1
                        description: Current page number
                      limit:
                        type: number
                        example: 10
                        description: Items per page
                      totalPages:
                        type: number
                        example: 10
                        description: Total number of pages
                    required:
                      - total
                      - page
                      - limit
                      - totalPages
                  error:
                    type: object
                    nullable: true
                    example: null
                    description: Error information (null on success)
                  errorCode:
                    type: string
                    nullable: true
                    example: null
                    description: Error code (null on success)
                required:
                  - statusCode
                  - message
                  - data
                  - paging
      security: *ref_4
      summary: Get orders by shop
      tags: *ref_5
...
info:
  title: API Docs
  description: Custom Swagger UI in NestJS
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    OrderItemResponseDto:
      type: object
      properties:
        productId:
          type: string
          example: prod_abc123
        productName:
          type: string
          example: K-Pop Album - Limited Edition
        quantity:
          type: number
          example: 2
        unitPrice:
          type: number
          example: 25000
        total:
          type: number
          example: 50000
      required:
        - productId
        - productName
        - quantity
        - unitPrice
        - total
    ShippingAddressResponseDto:
      type: object
      properties:
        recipientName:
          type: string
          example: Nguyen Van A
        phoneNumber:
          type: string
          example: '0912345678'
        addressLine:
          type: string
          example: 123 Main Street
        ward:
          type: string
          example: Ward 10
        district:
          type: string
          example: District 3
        province:
          type: string
          example: Ho Chi Minh City
        postalCode:
          type: string
          example: '700000'
        fullAddress:
          type: string
          example: 123 Main Street, Ward 10, District 3, Ho Chi Minh City, 700000
      required:
        - recipientName
        - phoneNumber
        - addressLine
        - ward
        - district
        - province
        - postalCode
        - fullAddress
    OrderResponseDto:
      type: object
      properties:
        id:
          type: string
          example: order_abc123
        customerId:
          type: string
          example: user_xyz789
        shopId:
          type: string
          example: shop_def456
        status:
          type: string
          example: PENDING_PAYMENT
          enum:
            - PENDING_PAYMENT
            - CONFIRMED
            - PAID
            - SHIPPING
            - DELIVERED
            - CANCELED
            - EXPIRED
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponseDto'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddressResponseDto'
        paymentMethod:
          type: string
          example: CREDIT
          enum:
            - CREDIT
            - COD
        paymentId:
          type: object
          example: payment_123456
          nullable: true
        trackingNumber:
          type: object
          example: TRACK12345
          nullable: true
        cancelReason:
          type: object
          example: Customer changed mind
          nullable: true
        totalAmount:
          type: number
          example: 75000
        createdAt:
          format: date-time
          type: string
          example: '2024-01-15T08:30:00Z'
        updatedAt:
          format: date-time
          type: string
          example: '2024-01-20T14:45:00Z'
      required:
        - id
        - customerId
        - shopId
        - status
        - items
        - shippingAddress
        - paymentMethod
        - paymentId
        - trackingNumber
        - cancelReason
        - totalAmount
        - createdAt
        - updatedAt
    CreateOrderDto:
      type: object
      properties: {}
    MarkOrderAsPaidDto:
      type: object
      properties:
        paymentId:
          type: string
          example: pay_123456789
      required:
        - paymentId
    ShipOrderDto:
      type: object
      properties:
        trackingNumber:
          type: string
          example: TRACK123456789
    CancelOrderDto:
      type: object
      properties:
        reason:
          type: string
          example: Customer changed mind
...